+++ neurolabi/gui/dvid/zdvidwriter.h
@@ -67,8 +67,8 @@ public:
 
   void createData(const std::string &type, const std::string &name);
 
-  void writeBodyInfo(int bodyId, const ZJsonObject &obj);
-  void writeBodyInfo(int bodyId);
+  void writeBodyInfo(uint64_t bodyId, const ZJsonObject &obj);
+  void writeBodyInfo(uint64_t bodyId);
   //void writeMaxBodyId(int bodyId);
 
   void mergeBody(const std::string &dataName, int targetId,
